[gdb] Allow frv::fr300 in selftests
In skip_arch in gdb/selftest-arch.c we skip architecture fr300 because of
PR20946, but the PR has been fixed by commit 6ca35dfbd1b ("Prevent an abort in
the FRV disassembler if the target bfd name is unknown.") in Januari 2017.

Remove the skipping of frv::fr300.

Tested on x86_64-linux.