PowerPC64: stub debug dump
powerpc64le-linux-ld is failing the assertion in ppc_build_one_stub,
again apparently, which means a stub will overwrite the tail of a
previous stub.  The difficulty with debugging these issues is that
it's not a problem with the stub that triggers the assertion, but the
previous stub in that section.  This patch keeps track of the last
stub and adds a debug dump.  Hopefully that will help.

	* elf64-ppc.c (enum _ppc64_sec_type): Add sec_stub.
	(struct _ppc64_elf_section_data): Add u.last_ent.
	(dump_previous_stub): New function.
	(ppc_build_one_stub): Keep track of previous stub, and dump it
	when finding an overlapping stub.